parser_website/requirements.txt
base91==1.0.1
beautifulsoup4==4.6.0
certifi==2017.4.17
chardet==3.0.4
click==6.7
Django==1.11.2
Flask==0.12.2
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
olefile==0.44
Pillow==4.1.1
pytz==2017.2
requests==2.17.3
urllib3==1.21.1
Werkzeug==0.12.2
